% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_hessisch}
\alias{get_hessisch}
\title{get_hessisch}
\usage{
get_hessisch(txt)
}
\arguments{
\item{txt}{provide a text you want to translate from German into Hessian}
}
\description{
Translates German to Hessian
}
